+++ bfv/keys.go
@@ -40,7 +40,7 @@ func NewSwitchingKey(params *Parameters) *SwitchingKey {
 }
 
 // NewRelinearizationKey returns an allocated BFV public relinearization key with zero value for each degree in [2 < maxRelinDegree].
-func NewRelinearizationKey(params *Parameters, maxRelinDegree uint64) *RelinearizationKey {
+func NewRelinearizationKey(params *Parameters, maxRelinDegree int) *RelinearizationKey {
 	return &RelinearizationKey{*rlwe.NewRelinKey(maxRelinDegree, params.N(), params.QPiCount(), params.Beta())}
 }
